Fix typos in CellMeasurer example modules
